requirements.txt
# Core dependencies for Gemini API and Ollama projects
google-generativeai>=0.3.0
ollama>=0.1.0
python-dotenv>=1.0.0  # For backward compatibility with .env files

# Jupyter and IPython dependencies for notebook functionality
jupyter>=1.0.0
ipython>=8.0.0
notebook>=7.0.0

# Streamlit for interactive testing app
streamlit>=1.28.0

# Additional dependencies that might be needed
requests>=2.31.0
pathlib2>=2.3.0

# Code formatting and debugging tools
black>=23.0.0

# Common data science libraries (optional - for testing)
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
